remove useSearchHandler and inline it in SearchBar
