+++ packages/nextjs/components/blockexplorer/SearchBar.tsx
@@ -1,7 +1,33 @@
-import { useSearchHandler } from "~~/hooks/scaffold-eth";
+import { useState } from "react";
+import { useRouter } from "next/router";
+import { ethers } from "ethers";
+import { localhost } from "wagmi/chains";
+import { getLocalProvider } from "~~/utils/scaffold-eth";
 
+const provider = getLocalProvider(localhost);
 export const SearchBar = () => {
-  const { handleSearch, searchInput, setSearchInput } = useSearchHandler();
+  const [searchInput, setSearchInput] = useState("");
+  const router = useRouter();
+
+  const handleSearch = async (event: React.FormEvent) => {
+    event.preventDefault();
+    if (ethers.utils.isHexString(searchInput)) {
+      try {
+        const tx = await provider?.getTransaction(searchInput);
+        if (tx) {
+          router.push(`/blockexplorer/transaction/${searchInput}`);
+          return;
+        }
+      } catch (error) {
+        console.error("Failed to fetch transaction:", error);
+      }
+    }
+
+    if (ethers.utils.isAddress(searchInput)) {
+      router.push(`/blockexplorer/address/${searchInput}`);
+      return;
+    }
+  };
 
   return (
     <form onSubmit={handleSearch} className="flex items-center justify-end mb-5 space-x-3">
